football-api/pyproject.toml
[tool.poetry]
name = "football-api"
version = "1.0.0"
description = "API REST para gestão de grupos e partidas de futebol"
authors = ["Your Name <you@example.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }
sqlalchemy = { version = "^2.0.0", extras = ["asyncio"] }
asyncpg = "^0.29.0"
alembic = "^1.13.0"
pydantic = { version = "^2.8.0", extras = ["email"] }
pydantic-settings = "^2.4.0"
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
bcrypt = "4.0.1"
python-multipart = "^0.0.12"
structlog = "^24.4.0"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
pytz = "^2024.1"
nanoid = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
asgi-lifespan = "^2.1.0"
ruff = "^0.6.0"
black = "^24.8.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
